+++ src/csurf/controls/csurf_valuebar.cpp
@@ -16,7 +16,7 @@ void CSurf_Valuebar::SendMode()
     }
 };
 
-CSurf_Valuebar::CSurf_Valuebar(int channel, double value, midi_Output *m_midiout) : value(value), m_midiout(m_midiout)
+CSurf_Valuebar::CSurf_Valuebar(int channel, int value, midi_Output *m_midiout) : value(value), m_midiout(m_midiout)
 {
     channelValuebar = channelValuebars[channel];
     channelValuebarMode = channelValuebarModes[channel];
